requirements.txt
# Discord bot framework
discord.py==2.4.0

# Python 3.13+ compatibility (audioop was removed in 3.13)
audioop-lts==0.2.1

# Google Gemini AI
google-generativeai==0.8.3

# Environment variable management
python-dotenv==1.0.1

